+++ CHANGELOG.md
@@ -9,7 +9,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/
 [Semantic Versioning]: https://semver.org/spec/v2.0.0.html
 
-## [Unreleased]
+## [0.10.0] - 2020-11-11
 
 ### Added
 
@@ -147,6 +147,7 @@ The format is based on [Keep a Changelog], and this project adheres to
 [0.7.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.7.0
 [0.8.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.8.0
 [0.9.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.9.0
+[0.10.0]: https://github.com/dogmatiq/dogma/releases/tag/v0.10.0
 
 <!-- version template
 ## [0.0.1] - YYYY-MM-DD